Config removal from jobdata
The jobdata feature preserves the job configuration after job execution.
Historically, a distinction existed between the configuration stored in the
'config' file and that stored in 'args.json' due to legacy features. However,
with the introduction of 'avocado.core.settings', the job configuration has
been consolidated into a single location, which caused the differentiation
between the 'config' file and 'args.json' is unnecessary. Consequently, we can
eliminate the 'config' file from the jobdata directory, as the same data is now
available in the 'args.json' file.

Reference: #5857
Signed-off-by: Jan Richter <jarichte@redhat.com>