avocado.Test: remove runner queue
The runner_queue property of avocado.Test (and related methods and
attributes) is used by the legacy runner only.

This is a change in the core test API, but as can be seen from the
fact that the latest LTS release (103.0) does not ship with the legacy
runner, it's an interface that is not used and produces no effect.
Thus, it can and should be removed.

Fixes: https://github.com/avocado-framework/avocado/issues/5340
Signed-off-by: Cleber Rosa <crosa@redhat.com>