Update json result attributes to match with job.result.tests
This is a fix for json result plugin, because the attributes in
`results.json` file and `job.result.tests` variable had different names
for the same data. This issue created confusion for users and also made
comparison of these data much harder. This commit changes the names in
json result plugin to correspond `job.result.tests`. Also, it adds
console warning for users, to be notified about this change.

Reference: #5826
Signed-off-by: Jan Richter <jarichte@redhat.com>