Introduce runnable recipe schema validation
The runnable recipe, that is, the representation of runnables in JSON
files, maps very close to the avocado.core.nrunner.runnable.Runnable
class, but there has been no actual verification of its structure.

This introduces a JSON schema file that should better define the
runnable recipe structure.  It als adds a test that checks all the
existing example runnable recipes.

Signed-off-by: Cleber Rosa <crosa@redhat.com>