Job dependencies
This commit introduces job dependency feature. With this feature, you can
define common dependency for each test inside job or with Job API inside suite.
With this feature, it should be easier to create tests which share their
dependencies.

It reads all the dependencies from the json file, where the user can specify
all the dependencies needed for this job. It uses the same json format as is
used in test docstrings. Part of this commit is a documentation update which
describes this feature more deeply.

Reference: #5504
Signed-off-by: Jan Richter <jarichte@redhat.com>