Introduce "Runnable Recipe" resolver
Avocado uses a Runnable to describe what is to be executed (usually a
test) in a given Task.  Runnables (and Tasks) can be defined in JSON
files, which are called Recipes (see the examples/nrunner/recipes
directory).

These can be loaded and executed with a command such as:

  avocado-runner-exec-test runnable-run-recipe \
     examples/nrunner/recipes/runnables/exec_test_echo_no_newline.json

It may also be useful to run them with an Avocado job.  To do that, a
"runnable-recipe" resolver is being introduced, which can read from
these JSON files and create runtime runnables (which will be added to
a job, a suite and turned into a runtime task).

Signed-off-by: Cleber Rosa <crosa@redhat.com>