Implement runtime validation of runnable recipes
Even though there's already checks for the example runnable recipe
files we ship, many bugs and weird behaviors can arise if we feed
broken or missing data to the resolvers, runners, etc.

This attempts to prevent such situations when the runnable recipe come
from an external source, let's say, when a user is writing a recipe by
hand.

Because we don't want to increase the dependencies on the Avocado
package, quite the contrary, we attempt to use the "jsonschema"
package when available, but, if it's not, then we perform a weaker but
still valid check.

For people using Avocado from RPM packages, because there are already
some dependencies being pulled, jsonschema becomes an additional
dependency.

Signed-off-by: Cleber Rosa <crosa@redhat.com>