avocado/core/output.py: use fstrings instead of addition of strings
First, there's a slight performance advantage in using fstrings
instead of adding strings together.  Second, this changes makes the
output utility not enforce policy when it comes to types, that is,
if the message to be printed is "None", it will be converted to a
string, instead of raising TypeError.

Signed-off-by: Cleber Rosa <crosa@redhat.com>