avocado/plugins/list.py: use a more accurate name for variable
The "cls" variable name is a legacy choice, based on the fact that the
old runner would always be dealing with some Python classes (even for
the SIMPLE tests and the like, a matching class inheriting from
avocado.Test would exist).

This name is confusing when reading the code and thinking about
Avocado's current core concepts.  Let's rename this to match the fact
that we're talking about a test *kind*, the name terminology used on
other places such as runnable recipe files.

Signed-off-by: Cleber Rosa <crosa@redhat.com>