exec-test disable env variable
This commit adds more control over the environment variables in
exec-tests for users. It brings a way how to disable env variable during
the test runtime so it won't be available to test script. Only thing
which user needs to do is to set the variable to `None` in test
`kwargs`.

Reference: #5889
Signed-off-by: Jan Richter <jarichte@redhat.com>