store_logging_stream_external test fix
The store_logging_stream_external test installs external package
matplotlib and checks with the package properly logs to the avocado
directories. Unfortunately, this test used hardcoded line number of
matplotlib. Because of this, the test started failing when the line was
changed. This commit fixes this test by using regex instead of hardcoded
string. After this change, the test should be protected against external
changes in matplotlib package.

Signed-off-by: Jan Richter <jarichte@redhat.com>