Show origin of resolution in verbose list
When using some of the new resolvers such as "runnable-recipe",
there's no longer a 1:1 relationship that the resolver and the "kind"
match.

Certain types of resolution may look fine from the perspective of the
original resolver, while it may be a broken or invalid reference to
a more apt (or canonical) resolver for its set "kind".

By listing the original resolver, users may be able to more easily
identify what the discrepancy is and where it's coming from.

Reference: https://github.com/avocado-framework/avocado/issues/5924
Signed-off-by: Cleber Rosa <crosa@redhat.com>