Hardcoded nrunner removal in TestSuite
This is a removal of left over of legacy code. When avocado supported
`legacy` and `nrunner` runners we also distinguished between resolver
resolution and test load. When the `legacy` runner has been removed, we
don't have this issue anymore, because resolvers are only supported
implementation of test resolution. Unfortunately, we kept validation of
supported runners during the test reference resolution with hardcoded
`nrunner`. This is wrong because `nrunner` is a `SuiteRunner` plugin,
and it can be replaced by any installed `SuiteRunner`.

Reference: #5954
Signed-off-by: Jan Richter <jarichte@redhat.com>