Fix use of job/suite configuration on runnables
The setting of configuration into runnables is currently being done at
the time where a runnable will be run (at the "nrunner" suite runner
plugin).

But, there are situations, such as when using the Job API, that this
is too late or not effective.

This change moves the setting of the configuration to the suite, which
is based on the job configuration, and thus already has all the
information needed to set the final version of the configuration.

Fixes: https://github.com/avocado-framework/avocado/issues/5963
Signed-off-by: Cleber Rosa <crosa@redhat.com>