Runnable: only set configuration used by runners
This reverts commit 94584fc60 ("Runnable: do not ignore the
configuration passed from recipes").

That change, because of the fact that every runnable uses
"runner.identifier_format", erroneously assumed that every
configuration should be passed to the runnables.  In fact, it was
decided long ago, and implemented after that, that only the
configuration that is used by runnables should be set on them.

Now, because "runner.indentifier_format" is used by all, this needs to
be taken into account.  In the near future, we may revisit the idea of
"runner.indentifier_format" being handled by the runnable as a
configuration item, and be transformed into a primary parameter of the
runnables.

Reference: https://github.com/avocado-framework/avocado/issues/5964
Signed-off-by: Cleber Rosa <crosa@redhat.com>