Deployment test: update to Fedora 40 and matching ansible
The COPR repos are building for the active Fedora, so let's bump that
to version 40.  While at it, we will be using a version of ansible
that has deprecated the "include" tag.  The error one gets is:

   ERROR! [DEPRECATED]: ansible.builtin.include has been removed. Use
   include_tasks or import_tasks instead. This feature was removed from
   ansible-core in a release after 2023-05-16. Please update your
   playbooks.

This updates the playbook to work properly under the matching ansible
version.

Reference: 44bcb711b1ff2b1de8f6534804dc3c621005f08a
Signed-off-by: Cleber Rosa <crosa@redhat.com>