Runnable and Suite configuration support
This change allows for the a runnable configuration and the suite and
default configuration to coexist with the correct behavior.

In short, if the suite has a configuration, it will become the new
default configuration for a runnable, while its own configuration will
not be touched.

Because the purpose of the default_config and config are different,
the checks for the values they may contain are different.  The
default_config should contain the exact keys as in the used config.
The config itself, on the other hand, may contain nothing, or some
values, but it should never be more than a subset of the config.

Note: there's an opportunity here for changing the data
structure used for the "config" attribute, one that looks
up the value in the "default_config" if it doesn't exist
in the actual "config".  But, given that the size of the
"default_config" will always be very small (given that
it's filtered to contain only the used configuration
by the runner) it seemed an unnecessary optimization
at this time.

Fixes: https://github.com/avocado-framework/avocado/issues/5998
Signed-off-by: Cleber Rosa <crosa@redhat.com>