Runnable: introduce default config
Right now, there's a conflict of interest on the "config" attribute
of a runnable: it serves as both a convenience holder of default
values from Avocado's configuration (default values set by the
settings API itself or in current configuration files), and the
actual configuration defined in the runnable API.

This introduces a different config holder, one with the default
configuration known at the time the class is istantiated.

At this point, it's a separate entity, but it will cooperate with
config, so that the config attribute will provide either the value
set in the runnable itself or in the default config.

Signed-off-by: Cleber Rosa <crosa@redhat.com>