Static checks: use lint check from avocado-static-checks
This switches to the script provided by the avocado-static-checks
repo, while using the local pylintrc configuration (set through the
use of "avocado-static-checks.conf".

In the near future, at least the "avocado/utils" directory needs to
made compliant with the standard pylint configuration provided by
avocado-static-checks, so that code can be easily moved to autils.

Fixes: https://github.com/avocado-framework/avocado/issues/5990
Signed-off-by: Cleber Rosa <crosa@redhat.com>