Resolvers: refactor common code that checks for executable files
The exec-test and tap resolvers share the same code that verifies the
condition of the reference given being an executable file.

This would be enough to consolidate them into a single method, but
there's going to be a third users of the same check, so there's more
reason for that yet.

Signed-off-by: Cleber Rosa <crosa@redhat.com>