exec-runnable-recipe resolver: add support for arguments
This adds support for used defined arguments to be passed while
running the executables that will generate the runnables recipe JSON
content.

It gives the opportunity for either calling the executables with a
particular option that will output the runnables (instead of other
action), or tweaking the type of runnables that will be generated.

Signed-off-by: Cleber Rosa <crosa@redhat.com>