Distro: add a functional test for Debian probes
While the standard "debian" container is not so easily suitable for
being used in the existing functional tests for distros that run an
Avocado job, the very standard "python" container is built on Debian
and provides Python (of course).

This should serve to avoid regressions to the distro probes for
Debian.

Signed-off-by: Cleber Rosa <crosa@redhat.com>