lint: introduce pylint configuration for utils and address W1203
In order to have "avocado/utils" in a way that can be easily
transported to the autils project, it's necessary to have the closest
lint configuration possible (ideally, the same).

This introduces a configuration file that is a copy of the existing
".pylintrc" file, but removes the waive on W1203.  The goal is that
further commits will address other warnings and bring the exceptions
as close as possible to the "avocado-static-checks" default pylint
configuration file, which has no exceptions.

For this change, there was the possibility of making changes
compatible with "logging-format-style=old" (the default), or jumping
the gun and making changes compatible with "logging-format-style=new".
The earlier approach was chosen because the goal is to get closer to:

* A default pylint configuration (as shipped/generated by pylint
  itself)
* The configuration provided by avocado-static-checks and used by
  autils

Reference: https://pylint.pycqa.org/en/latest/user_guide/messages/warning/logging-fstring-interpolation.html
Signed-off-by: Cleber Rosa <crosa@redhat.com>