lint: address fixme / W0511
There are a few occurrences where code is being used to keep track of
possible improvements.  Many times these will go unnoticed for ages.

Let's either document the limitations that are affected by the
"TODO"s, or remove them because after quite some time, they were not
something that was attempted to be improved or fixed.

https://pylint.pycqa.org/en/latest/user_guide/messages/warning/fixme.html
Signed-off-by: Cleber Rosa <crosa@redhat.com>