Remove IPv4 Addresses of network interface.
when IP addresses are confiugred from a config file, the generic
"ip flush" command does not removes the IP address and IP's will
be remain intact.

when user wanted to remove all IP addresses of interface which user
doesnot aware of it, this code will helps to find all the assigned
IP addresses and removes it.

Signed-off-by: Shaik Abdulla <abdulla1@linux.vnet.ibm.com>