lint: introduce pylint configuration for utils
In order to have "avocado/utils" in a way that can be easily
transported to the autils project, it's necessary to have the closest
lint configuration possible (ideally, the same).

This introduces a configuration file that is a copy of the existing
".pylintrc" file.  The goal is that further commits will address other
warnings and bring the exceptions as close as possible to the
"avocado-static-checks" default pylint configuration file, which has
no exceptions.

Signed-off-by: Cleber Rosa <crosa@redhat.com>