Utils / Network: minimize failires of ping_flood()
The implementation of ping_flood() counts on the output generated by
the ping command line utility.  But, the way and amount of data it
reads looking for the failure pattern is both too small and is
susceptible to timing issues.

In order to minimize those possibilities, let's send error
messages (which will never contain the info this function is looking
for) to another file.  It also allows to use that as an improvement to
the log messages.

Signed-off-by: Cleber Rosa <crosa@redhat.com>