avocado-instrumented/python-unittest resolvers: improve fail message
The message given by the avocado-instrumented and python-unittest
resolvers, resolvers that share a lot of code and require Python
files, can be made clearer.

One of the earliest checks performed is about the file *name* and
its *suffix*, and not necessarily about an actual file (that exists).
Even though the concepts are pretty close, let's make it clear that
these resolvers are rejecting the references given based on the ".py"
suffix not existing in the file name, and not necessarily, at that
time, rejecting it based on an actual existing file.

Signed-off-by: Cleber Rosa <crosa@redhat.com>