selftests: drop job-api test variant that may exceed sqlite3's own limits
The sqlite3 module (and actually, even the non-python binary with the
same name), have "interesting" limitations when it comes to the
database path.

When the test that creates the journal file, and checks for it, runs
as part of an RPM build, the extra directories (/buildir/...) exceed
the path size limit.

While it would be possible to check if the "logdir" path, plus the
name of the journal database path, would eventually exceed the sqlite3
limits, this logic would be exclusive for one test variant.

IMO it does not justify adding such logic to keep a test that has
considerably low value.

Signed-off-by: Cleber Rosa <crosa@redhat.com>