Runners refactoring
This is a refactoring of runners. It moves most of the code related to
running and monitoring processes into base class. This change removes
duplicities across the runners, and it will help with implementation of
common features like signal handling. Also it makes the future
development of new runners easier, because developers can focus on
actual runner and don't have to deal with things like process
monitoring, logging and exception handling

Signed-off-by: Jan Richter <jarichte@redhat.com>