Coverage.py during test load
This commit updates the coverage.py support in avocado. Currently,
avocado was only covering lines which has been touched during the test
runtime, but such solution omits lines which are used during imports.
Because of this the coverage can be inaccurate, because it misses lines
with method names, class names, imports etc. After this change the
coverage reports are more accurate.

Signed-off-by: Jan Richter <jarichte@redhat.com>