Python publishing fails with InvalidDistribution Error
Based on the (pypa/twine#1216)
the root cause seems to be `setuptools`, but the workaround seems to
depend on the installed version of `packaging`.

`twine` declares a dependency on `packaging>=24.0.0` and the bug doesn't
appear if `24.2` is installed. As a result, the problem shows on Python
distributions that come with `packaging@24.0` or `packaging@24.1`
preinstalled, which is used in CI. Therefore, we need to declare
`packaging>=24.2` requirement during twine installation.

Reference: https://github.com/pypa/twine/issues/1216
Signed-off-by: Jan Richter <jarichte@redhat.com>