Github workflow for announcement of AAutils migration
This workflow should notify contributors on their PRs when they are
changing an already migrated utility. This should help us to avoid
conflicts between avocado.utils and aautils. The announcement looks like
this:

```
You are changing one of the avocado utils which has already been
migrated to AAutils project https://github.com/avocado-framework/aautils
and this utility will be removed after the LTS release. Please make sure
that all your proposed changes are already in AAutils and this PR is
only backport.

For more information about AAutlis migration see
https://avocado-framework.readthedocs.io/en/latest/blueprints/BP005.html

For a list of migrated utilities see
https://avocado-framework.github.io/aautils.html
```

Signed-off-by: Jan Richter <jarichte@redhat.com>