Selftests / Functional / Runners: simplify test class inheritance
There's no need to use multiple inheritance in these test classes.

The choices made here was to use the simplest necessary class, which
resulted in one being TestCaseTmpDir (basically a unittest.TestCase
with a few extra methods for temporary dirs), and the other an
Avocado.Test class because of the use of self.log().

Signed-off-by: Cleber Rosa <crosa@redhat.com>