Asset tests: convert to functional tests
The current version of the tests patch the Asset.fetch() method to
simulate a successful or failed fetch.  The problem is that mocks are
not guaranteed across processes, and the AssetRunner spawns a separate
process to fetch the asset without blocking.

This changes the test so that an actual fetch is performed, without
mocking.

Fixes: https://github.com/avocado-framework/avocado/issues/6190
Signed-off-by: Cleber Rosa <crosa@redhat.com>