GDB/MI parser regression fix
The first member of "t_{type_of_token}" methods are used as the regex
to match while scanning the input.  With the previous change, no
symbol would be caught because the effective regex would not be
considered.

This reverts 2d741f68909c7ae736830f6a76fd7c88554fa179 and adds a test
to avoid future breakage.

Signed-off-by: Cleber Rosa <crosa@redhat.com>