Docs: ignore external libraries used by utilities
The "avocado.utils.external" namespace is not intended to be used
directly by Avocado tests, and in the migration to AAutils, it will be
put in a different namespace (if needed) to avoid further confusion.

This also works around an issue with Sphinx (docutils) trying to
substitute references in a regex as if it were actual reStructuredText.

Signed-off-by: Cleber Rosa <crosa@redhat.com>