Improve reliability of flaky test test_task_status_service_lost
The TaskRunStatusService.test_task_status_service_lost, from
selftests/functional/nrunner.py, fails every now and then.  There are
a number of expectations with regards to the timing of both the
netcat process and listening socket, and the avocado-runner process.

The following changes are added here in an effort to mitigate or
eliminate those failures:

 * A wait was added after the netcat process starts, and before the
   avocado-runner process is created and started, improving the odds
   that the netcat socket will be available.

 * A longer execution time for the underlying executable (/bin/sleep)
   mitigates the possibility of it ending before netcat is killed.

 * A longer sleep after the avocado-runner process is started improves
   the odds of it having properly started and connected to the socket.

 * Waiting for the end of the netcat process makes sure the assertion
   is not tried without the connection having being broken on the
   netcat side.

Signed-off-by: Cleber Rosa <crosa@redhat.com>